error finding
while looking for error, a 4 and 5 that are  invisible in the text file are found. since the issue arouse from the program reading the file, but the code of the program itself doesn't seem to have issues. the code of reading the txt file will no longer be edited but data will be changed within code.